+++ MapboxSearch/offline/src/main/java/com/mapbox/search/offline/OfflineSearchEngineImpl.kt
@@ -66,6 +66,11 @@ internal class OfflineSearchEngineImpl(
         coreEngine.selectTileset(dataset, version)
     }
 
+    @MapboxExperimental
+    override fun selectTileset(tilesetParameters: TilesetParameters) {
+        coreEngine.selectTileset(tilesetParameters.generatedDatasetName, tilesetParameters.version)
+    }
+
     override fun search(
         query: String,
         options: OfflineSearchOptions,