+++ MapboxSearch/sample/src/androidTest/java/com/mapbox/search/sample/ExampleActivitiesTest.kt
@@ -22,6 +22,7 @@ import com.mapbox.search.sample.api.HistoryDataProviderKotlinExample
 import com.mapbox.search.sample.api.JapanSearchJavaExampleActivity
 import com.mapbox.search.sample.api.JapanSearchKotlinExampleActivity
 import com.mapbox.search.sample.api.OfflineEvSearchKotlinExampleActivity
+import com.mapbox.search.sample.api.OfflineMultipleLanguagesActivity
 import com.mapbox.search.sample.api.OfflineReverseGeocodingJavaExampleActivity
 import com.mapbox.search.sample.api.OfflineReverseGeocodingKotlinExampleActivity
 import com.mapbox.search.sample.api.OfflineSearchJavaExampleActivity
@@ -90,6 +91,8 @@ class ExampleActivitiesTest(private val clazz: Class<out Activity>) {
             OfflineReverseGeocodingJavaExampleActivity::class.java,
 
             OfflineEvSearchKotlinExampleActivity::class.java,
+
+            OfflineMultipleLanguagesActivity::class.java,
 
             JapanSearchKotlinExampleActivity::class.java,
             JapanSearchJavaExampleActivity::class.java,