+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/MainActivity.kt
@@ -69,6 +69,7 @@ import com.mapbox.search.sample.api.HistoryDataProviderKotlinExample
 import com.mapbox.search.sample.api.JapanSearchJavaExampleActivity
 import com.mapbox.search.sample.api.JapanSearchKotlinExampleActivity
 import com.mapbox.search.sample.api.OfflineEvSearchKotlinExampleActivity
+import com.mapbox.search.sample.api.OfflineMultipleLanguagesActivity
 import com.mapbox.search.sample.api.OfflineReverseGeocodingJavaExampleActivity
 import com.mapbox.search.sample.api.OfflineReverseGeocodingKotlinExampleActivity
 import com.mapbox.search.sample.api.OfflineSearchAlongRouteExampleActivity
@@ -106,7 +107,7 @@ class MainActivity : AppCompatActivity() {
 
     private lateinit var tileRegionLoadOptions: TileRegionLoadOptions
     private val tileStore = TileStore.create()
-    private val tileRegionId = "Washington DC"
+    private val tileRegionId = "MainActivity - Washington DC"
     private var tilesLoadingTask: Cancelable? = null
 
     private val onBackPressedCallback = object : OnBackPressedCallback(false) {
@@ -123,7 +124,7 @@ class MainActivity : AppCompatActivity() {
                     if (BuildConfig.DEBUG) {
                         error("This OnBackPressedCallback should not be enabled")
                     }
-                    Log.i("SearchApiExample", "This OnBackPressedCallback should not be enabled")
+                    Log.i(LOG_TAG, "This OnBackPressedCallback should not be enabled")
                     isEnabled = false
                     onBackPressedDispatcher.onBackPressed()
                 }
@@ -225,12 +226,17 @@ class MainActivity : AppCompatActivity() {
         tilesLoadingTask = tileStore.loadTileRegion(
             tileRegionId,
             tileRegionLoadOptions,
-            { progress -> Log.i("SearchApiExample", "Loading progress: $progress") },
+            { progress -> Log.i(LOG_TAG, "Loading progress: $progress") },
             { result ->
-                if (result.isValue) {
-                    Log.i("SearchApiExample", "Tiles successfully loaded: ${result.value}")
+                val loadResultMessage = if (result.isValue) {
+                    "Tiles successfully loaded: ${result.value}"
                 } else {
-                    Log.i("SearchApiExample", "Tiles loading error: ${result.error}")
+                    "Tiles loading error: ${result.error}"
+                }
+
+                Log.i(LOG_TAG, loadResultMessage)
+                runOnUiThread {
+                    Toast.makeText(applicationContext, loadResultMessage, Toast.LENGTH_SHORT).show()
                 }
             }
         )
@@ -520,6 +526,10 @@ class MainActivity : AppCompatActivity() {
                 startActivity(Intent(this, OfflineEvSearchKotlinExampleActivity::class.java))
                 true
             }
+            R.id.open_offline_multiple_languages_kt_example -> {
+                startActivity(Intent(this, OfflineMultipleLanguagesActivity::class.java))
+                true
+            }
             R.id.open_history_data_provider_java_example -> {
                 startActivity(Intent(this, HistoryDataProviderJavaExample::class.java))
                 true
@@ -618,6 +628,8 @@ class MainActivity : AppCompatActivity() {
     }
 
     private companion object {
+
+        const val LOG_TAG = "MainActivity"
 
         val MARKERS_EDGE_OFFSET = dpToPx(64f).toDouble()
         val PLACE_CARD_HEIGHT = dpToPx(300f).toDouble()